pyproject.toml
[tool.poetry]
name = "django-enduro-mcb"
version = "0.1.0"
description = ""
authors = ["Heiden <119899341+CarlosHeiden@users.noreply.github.com>"]
readme = "README.md"
package-mode = false  # <<< aqui desabilita o empacotamento

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
django-import-export = "^4.3.9"
mysqlclient = "^2.2.7"
weasyprint = "^66.0"
xhtml2pdf = "^0.2.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
